Added app navigation and many new components
