+++ vibecraft/components/homepage.js
@@ -1,0 +1,14 @@
+// Homepage.js
+import React from 'react';
+import { View, Text } from 'react-native';
+
+const Homepage = () => {
+  return (
+    <View>
+      <Text>Welcome to Your App!</Text>
+      {/* Add more content as needed */}
+    </View>
+  );
+};
+
+export default Homepage;
