+++ vibecraft/package.json
@@ -10,6 +10,8 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^19.0.0",
+    "@react-navigation/native": "^6.1.9",
+    "@react-navigation/stack": "^6.3.20",
     "expo": "~49.0.15",
     "expo-crypto": "^12.6.0",
     "expo-linear-gradient": "~12.3.0",
